Consistent ranking points as float
git-svn-id: https://localhost/svn/wxTTM/trunk@7729 7754d8ab-9b54-e011-a251-0001c006bdc7

+++ src/wxTTM/Draw/DrawInf.h
@@ -121,10 +121,10 @@ class  DrawItemTeam : public DrawItem
     int   lastPos;            // Plazierung in Gruppe
     
     // Qualification + CP
-    bool  rkDirectEntry;      // Qualifier / Direct Entry
-    int   rkNatlRank;         // Nationales ranking
-    int   rkIntlRank;         // Internationalies Ranking
-    int   rankPts;            // Ranking points
+    bool   rkDirectEntry;      // Qualifier / Direct Entry
+    int    rkNatlRank;         // Nationales ranking
+    int    rkIntlRank;         // Internationalies Ranking
+    double rankPts;            // Ranking points
     
     bool  mask;
